Refactor out some common functionality to a shared utils class
